Update workoutday.cpp
Added user name, age, and weight to make the system more personal. System will also ask if weight has. You can input a new weight and it will appear in the header. Would like to update to input name, age, and weight of other users to make system more personal to others besides myself. Will work on that in the future