Updated code for work out reminder system
Added a theme header to the system. separated code into 3 different files for better readability. Also was able to add input from user so that system can record user data. User will be able to see which workout they should be performing that day as well as, advice on rest days and if user has not worked out.